requirements.txt
flask
Pillow
pystray
waitress
win32; sys_platform == 'win32'